+++ packages/typespec-rust/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-* Fixed the `allowEmpty` `@clientOption` decorator to get applied when it comes from the client spec, and not only from the routes.
+* Changed the `allowEmpty` `@clientOption` decorator to `minLength`.
 
 ## 0.39.0 (2026-04-10)